Make DataFutures complete after stageout (issue #778)
Prior to this, DataFutures were marked as complete when the main app
task completed, prior to any staging happening.

This meant that DataFutures completing were not an indication that
Globus stage-out had completed (or even been started).

This commit causes DataFutures to complete when the relevant
staging task has completed.

Prior to this commit, DataFutures were created by bash/python
app decorators. This commit creates the DataFutures as part
of stageout setup.

When stageout is inhibited, we still need to create the
DataFutures; in such case, a DataFuture will be marked as complete
as soon as its app future is complete (the existing behaviour)
- this is because the idea is that a stageout task completing
means that the outputs are done.

Fixes #778