Update EC2 test configuration to use new staging mechanisms
This means that EC2 tests that use http and file URLs will
now (maybe?) work. The tests that use globus and/or ftp
URLs will not.

The import statement for user_opts is modified so that the
modified config file can be used in the --config parameter
to pytest - previously, relative imports did not work in
that case.